requirements/base.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beanie==1.22.6
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
lazy-model==0.2.0
mongomock==4.1.2
mongomock-motor==0.0.21
motor==3.3.1
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.5.0
pytest
pytest-mock==3.11.1
sentinels==1.0.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
pydash==5.0.0
web3==6.0.0
sentry-sdk==1.12.1
celery[amqp]==5.3.1
redis==4.5.5
orjson==3.8.4
pydantic-settings==2.1.0
base58==2.1.1
PyNaCl==1.5.0